+++ Source/Core/InputCommon/CMakeLists.txt
@@ -2,6 +2,7 @@ add_library(inputcommon
   InputConfig.cpp
   InputProfile.cpp
   ControllerEmu/ControllerEmu.cpp
+  ControllerEmu/StickGate.cpp
   ControllerEmu/Control/Control.cpp
   ControllerEmu/Control/Input.cpp
   ControllerEmu/Control/Output.cpp
